+++ edx_arch_experiments/apps.py
@@ -3,7 +3,7 @@ edx_arch_experiments Django application initialization.
 """
 
 from django.apps import AppConfig
-from edx_django_utils.plugins.constants import PluginSettings
+from edx_django_utils.plugins.constants import PluginSettings, PluginURLs
 
 
 class EdxArchExperimentsConfig(AppConfig):
@@ -22,5 +22,12 @@ class EdxArchExperimentsConfig(AppConfig):
                     PluginSettings.RELATIVE_PATH: 'settings.production',
                 },
             }
-        }
+        },
+        PluginURLs.CONFIG: {
+            'lms.djangoapp': {
+                PluginURLs.NAMESPACE: 'edx_arch_experiments',
+                PluginURLs.REGEX: r'^api/certificates/v1/',
+                PluginURLs.RELATIVE_PATH: 'certificates.urls',
+            }
+        },
     }